fitness-ai-coach-backend/pyproject.toml
[project]
name = "fitness-ai-coach-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "3.11.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",  
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "router (>=0.1,<0.2)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langchain (>=0.3.17,<0.4.0)",
    "langchain-google-genai (>=2.0.9,<3.0.0)",
    "grpcio (>=1.60.0,<2.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
